rename AssetEnvelope & TrackOptions & add fade multiplier
