+++ src/mixer/AssetEnvelope.js
